+++ packages/squeak/next.config.js
@@ -1,3 +1,5 @@
+const removeImports = require('next-remove-imports')()
+
 const { withSentryConfig } = require('@sentry/nextjs')
 
 /** @type {import('next').NextConfig} */
@@ -29,4 +31,4 @@ const sentryWebpackPluginOptions = {
 
 // Make sure adding Sentry options is the last code to run before exporting, to
 // ensure that your source maps include changes from all other Webpack plugins
-module.exports = withSentryConfig(moduleExports, sentryWebpackPluginOptions)
+module.exports = removeImports(withSentryConfig(moduleExports, sentryWebpackPluginOptions))
