+++ packages/squeak/package.json
@@ -28,6 +28,8 @@
         "@types/passport-jwt": "^3.0.6",
         "@types/passport-local": "^1.0.34",
         "@types/passport-oauth2": "^1.4.11",
+        "@uiw/react-markdown-preview": "^4.1.5",
+        "@uiw/react-md-editor": "^3.19.7",
         "ajv": "^8.11.0",
         "babel-plugin-superjson-next": "^0.4.3",
         "bcrypt": "^5.0.1",
@@ -48,6 +50,7 @@
         "next": "^12.3.1",
         "next-absolute-url": "^1.2.2",
         "next-connect": "^0.12.2",
+        "next-remove-imports": "^1.0.8",
         "nextjs-cors": "^2.1.0",
         "node-pg-migrate": "^6.2.2",
         "nookies": "^2.5.2",
