+++ packages/squeak/src/pages/team/[id].tsx
@@ -18,6 +18,7 @@ import _ from 'lodash'
 import { GetTeamResponse } from '../../pages/api/teams'
 import ImageUpload from 'src/components/ImageUpload'
 import { SqueakConfig } from '@prisma/client'
+import { RoadmapForm } from 'src/components/RoadmapForm'
 
 const AddTeamMember = ({ teamId, onSubmit }) => {
     const [profiles, setProfiles] = useState<GetProfilesResponse[] | null | undefined>(null)
@@ -79,180 +80,6 @@ const AddTeamMember = ({ teamId, onSubmit }) => {
             <Button className="mt-3">Add</Button>
         </form>
     ) : null
-}
-
-export const RoadmapForm = ({ onSubmit, handleDelete, initialValues, submitText = 'Add goal', categories, config }) => {
-    const [confirmDelete, setConfirmDelete] = useState(false)
-    const [query, setQuery] = useState('')
-    const filteredCategories =
-        query === ''
-            ? categories
-            : categories.filter((category) => {
-                  return category.toLowerCase().includes(query.toLowerCase())
-              })
-    return (
-        <Formik initialValues={initialValues} onSubmit={onSubmit}>
-            {({ values, setFieldValue }) => {
-                return (
-                    <Form>
-                        <div className="flex space-x-6">
-                            <div className="flex-grow">
-                                <Input value={values.title} label="Title" id="title" name="title" placeholder="Title" />
-                                <Input
-                                    id="description"
-                                    value={values.description}
-                                    as={'textarea'}
-                                    label="Description"
-                                    name="description"
-                                    placeholder="Description"
-                                />
-
-                                {values.complete && (
-                                    <Input
-                                        value={values.date_completed}
-                                        label="Date completed"
-                                        id="date_completed"
-                                        name="date_completed"
-                                        placeholder="Date completed"
-                                        type="date"
-                                    />
-                                )}
-                                {!values.complete && (
-                                    <Input
-                                        value={values.projected_completion_date}
-                                        label="Projected completion date"
-                                        id="projected_completion_date"
-                                        name="projected_completion_date"
-                                        placeholder="Projected completion date"
-                                        type="date"
-                                    />
-                                )}
-                                <div className="mb-6">
-                                    <label className="text-[16px] font-semibold opacity-75 my-2" htmlFor={'category'}>
-                                        Category
-                                    </label>
-                                    <Combobox
-                                        value={values.category}
-                                        onChange={(category) => setFieldValue('category', category)}
-                                    >
-                                        <div className="relative">
-                                            <Combobox.Input
-                                                id="category"
-                                                className="block px-4 py-2 pr-0 border-gray-light border rounded-md w-full"
-                                                onChange={(event) => {
-                                                    setFieldValue('category', event.target.value)
-                                                    setQuery(event.target.value)
-                                                }}
-                                            />
-                                            <Combobox.Button className="absolute inset-y-0 right-0 flex items-center rounded-r-md px-2 focus:outline-none">
-                                                <ChevronDownIcon className="h-5 w-5 text-gray-400" aria-hidden="true" />
-                                            </Combobox.Button>
-                                        </div>
-                                        <Combobox.Options className="shadow-md rounded-md">
-                                            {filteredCategories.map((category) => (
-                                                <Combobox.Option
-                                                    className="cursor-pointer m-0 py-3 px-2 "
-                                                    key={category}
-                                                    value={category}
-                                                >
-                                                    {category}
-                                                </Combobox.Option>
-                                            ))}
-                                        </Combobox.Options>
-                                    </Combobox>
-                                </div>
-
-                                {values.github_urls.map((issue, index) => {
-                                    return (
-                                        <div key={index} className="flex items-center space-x-2">
-                                            <div className="flex-grow">
-                                                <Input
-                                                    id="github-url"
-                                                    value={values.github_urls[index]}
-                                                    label="GitHub URL"
-                                                    name={`github_urls[${index}]`}
-                                                    placeholder="GitHub URL"
-                                                />
-                                            </div>
-                                            <button
-                                                type="button"
-                                                onClick={() => {
-                                                    const issues = [...values.github_urls]
-                                                    issues.splice(index, 1)
-                                                    setFieldValue('github_urls', issues)
-                                                }}
-                                            >
-                                                <XIcon className="w-4 text-[red]" />
-                                            </button>
-                                        </div>
-                                    )
-                                })}
-                                <button
-                                    type="button"
-                                    onClick={() => setFieldValue('github_urls', [...values.github_urls, ''])}
-                                    className="text-red font-bold mb-6"
-                                >
-                                    Add GitHub URL
-                                </button>
-                            </div>
-                            <div className="w-48">
-                                {config.cloudinary_cloud_name && (
-                                    <ImageUpload
-                                        onChange={(file) => setFieldValue('image', file)}
-                                        value={values.image}
-                                    />
-                                )}
-                                <div className="flex items-center space-x-2 my-6">
-                                    <label
-                                        className="text-[16px] font-semibold opacity-75 flex-shrink-0 m-0"
-                                        htmlFor="complete"
-                                    >
-                                        Complete
-                                    </label>
-                                    <Field type="checkbox" name="complete" id="complete" />
-                                </div>
-                                <div className="flex flex-col">
-                                    <div className="flex items-center space-x-2 mb-6">
-                                        <label
-                                            className="text-[16px] font-semibold opacity-75 flex-shrink-0 m-0"
-                                            htmlFor="milestone"
-                                        >
-                                            Milestone
-                                        </label>
-                                        <Field type="checkbox" name="milestone" id="milestone" />
-                                    </div>
-
-                                    <div className="flex items-center space-x-2 mb-6">
-                                        <label
-                                            className="text-[16px] font-semibold opacity-75 flex-shrink-0 m-0"
-                                            htmlFor="beta_available"
-                                        >
-                                            Beta available
-                                        </label>
-                                        <Field type="checkbox" name="beta_available" id="beta_available" />
-                                    </div>
-                                </div>
-                            </div>
-                        </div>
-                        <div className="flex items-center mt-4 space-x-2 justify-between">
-                            {handleDelete && (
-                                <Button
-                                    onClick={() => {
-                                        if (!confirmDelete) return setConfirmDelete(true)
-                                        handleDelete()
-                                    }}
-                                    type="button"
-                                >
-                                    {confirmDelete ? 'Click to confirm' : 'Delete'}
-                                </Button>
-                            )}
-                            <Button>{submitText}</Button>
-                        </div>
-                    </Form>
-                )
-            }}
-        </Formik>
-    )
 }
 
 const Team = ({ id }) => {
